+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Go: Update dependencies.
+- Client: Some improvements. ([#619](https://github.com/giantswarm/clustertest/pull/619))
+  - Conditions: Rework `IsClusterReadyCondition`.
 
 ## [3.1.2] - 2026-02-03