+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Client: Rework `NewFromSecret`.
   - Client: Rework `GetClusterKubeConfig`.
   - Client: Rework `getTeleportKubeConfig`.
+  - Client: Rework `getCAPIKubeConfig`.
 
 ## [3.1.2] - 2026-02-03