+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.3] - 2026-02-17
+
 ### Changed
 
 - Go: Update dependencies.
@@ -904,7 +906,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial framework layout
 - Added Kubernetes client extended from controller-runtime client
 
-[Unreleased]: https://github.com/giantswarm/clustertest/compare/v3.1.2...HEAD
+[Unreleased]: https://github.com/giantswarm/clustertest/compare/v3.1.3...HEAD
+[3.1.3]: https://github.com/giantswarm/clustertest/compare/v3.1.2...v3.1.3
 [3.1.2]: https://github.com/giantswarm/clustertest/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/giantswarm/clustertest/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/giantswarm/clustertest/compare/v3.0.0...v3.1.0